LayoutTests/imported/w3c/indexeddb/idbversionchangeevent-expected.txt


FAIL IDBVersionChangeEvent fired in upgradeneeded, versionchange and deleteDatabase assert_equals: new version (versionchange) expected (object) null but got (number) 0
